Phase 2: GNU wc port (stdio/getopt/error/stat) + parity, kernel spawn fix
Port GNU textutils 2.1 wc end-to-end onto the Phase 2 libc and prove
byte-exact parity against both real GNU wc and the in-OS expectation
on bare metal.

libc (Phase 2 additions):
  - stdio.c: vsnprintf family with glibc-exact flags/width/precision
    (host-verified: 993,826 checks, ASan-clean)
  - getopt.c: POSIX getopt + GNU getopt_long (33 host scenarios vs glibc)
  - error.c: GNU error()/error_at_line()
  - stat.c: stat/fstat/lseek stubs (ENOSYS, graceful wc degradation)
  - file.c: buffered-read FILE layer; stdin/stdout/stderr statically
    bound to fds 0/1/2 (zeroed-bss FILE wrote stdout to fd 0)
  - exit(): weak fflush(0) so buffered output survives process exit
  - sysroot headers: stdio.h, unistd.h, fcntl.h, sys/types.h, sys/stat.h,
    getopt.h, error.h; stdlib/string/ctype from Phase 1

wc port:
  - src/user/wc.c: faithful textutils-2.1 wc (GPLv2), byte-identical
    layout, argv[0]-driven error prefixes; built for ARM + x64
  - src/user/wc_test.c: in-OS E2E spawning WC.BIN via spawn2/pipe,
    6 checks (file, stdin '-', multi-file total, long options) — all
    pass byte-exact on ARM (MT_EXIT=0, full suite halts)
  - src/host/wc_parity.sh + build_tu21_wc_ref.sh: strict parity vs a
    genuine textutils-2.1 build: 114/114 byte-exact matches

kernel fixes surfaced by the port:
  - sys_spawn: release the caller with -1 instead of parking it in
    WAIT_SPAWN forever when process_create_kernel cannot get a slot
    (was an unbounded system hang under process-table pressure) [both arches]
  - errno_test: tolerate EEXIST on the first mkdir since fat16_unlink
    refuses directories and there is no rmdir yet (x64 suite determinism)

Gates: make test (ARM) green incl. wc E2E; make unit_tests_intel (x64)
green; host_tests green; wc_parity_strict 114/114.

+++ src/host/files_test.c
@@ -80,7 +80,7 @@ void mock_read_dir_reset(void);
 void mock_sysinfo_override_reset(void);
 
 /* compat.c fd wrappers, used by the fd-1 capture helper below */
-int ho_read(int fd, void *buf, int size);
+ssize_t ho_read(int fd, void *buf, size_t size);
 int ho_close(int fd);
 int ho_pipe(int fds[2]);
